backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
web3==6.5.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.6.1
joblib==1.3.2
